fix build error

+++ pkg/square/square_test.go
@@ -375,7 +375,7 @@ func TestSquareShareCommitments(t *testing.T) {
 	dataSquare, err := builder.Export()
 	require.NoError(t, err)
 
-	cacher := inclusion.NewSubtreeCacher(dataSquare.Size())
+	cacher := inclusion.NewSubtreeCacher(uint64(dataSquare.Size()))
 	eds, err := rsmt2d.ComputeExtendedDataSquare(shares.ToBytes(dataSquare), appconsts.DefaultCodec(), cacher.Constructor)
 	require.NoError(t, err)
 	dah := da.NewDataAvailabilityHeader(eds)
